Изменил валидацию и сортировку
